Fix truncated binuint option value
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@1452 98333e67-4a24-44d7-a75c-e53540dd3050

+++ trunk/lib/util/cmdline_parser.c
@@ -500,7 +500,7 @@ cmd_getOptionValueBinUint(cmdlineParser const cpP,
 
     struct optionDesc * const optionDescP = findOptionDesc(cpP, name);
 
-    unsigned int retval;
+    uint64_t retval;
 
     if (!optionDescP) {
         fprintf(stderr, "cmdlineParser called incorrectly.  "
